scripts/pyproject.toml
[project]
name = "scripts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "python-dotenv>=1.0.1",
  "requests>=2.32.3",
  "supabase>=2.13.0",
  "tqdm>=4.67.1",
]
